updatesystem.c
The UpdateSystem program is a C script that automates the system update process, including package updates, full system upgrade, installation of the "neofetch" utility to display information, and removal of unnecessary packages. It streamlines the execution of these tasks with a single command, enhancing the efficiency and convenience of daily system updates.